:recycle: refacto import of utils
