+++ test/idAsNumber/active.test.ts
@@ -1,7 +1,11 @@
 import useUserStore from '../store/userStore'
-import { isArray, isArrayOfNumbers } from '../utils/array'
-import { defaultTestInit } from '../utils/baseState'
-import { getExpectedObjectProperties, user } from '../utils/dataFixtures'
+import {
+  defaultTestInit,
+  getExpectedObjectProperties,
+  isArray,
+  isArrayOfNumbers,
+  user,
+} from '../utils'
 
 describe('setActive action should return correct value', () => {
   beforeEach(() => {
