+++ test/idAsNumber/createOne.test.ts
@@ -1,8 +1,16 @@
 import { hasOwnProperty } from '@antfu/utils'
 import useUserStore from '../store/userStore'
-import { isArray, isArrayOfNumbers } from '../utils/array'
-import { defaultTestInit } from '../utils/baseState'
-import { getExpectedObjectProperties, user, user2, user3, user4, usersArray } from '../utils/dataFixtures'
+import {
+  defaultTestInit,
+  getExpectedObjectProperties,
+  isArray,
+  isArrayOfNumbers,
+  user,
+  user2,
+  user3,
+  user4,
+  usersArray,
+} from '../utils'
 
 describe('create action should return correct value', () => {
   beforeEach(() => {
