+++ test/idAsNumber/setCurrent.test.ts
@@ -1,8 +1,6 @@
 import { hasOwnProperty } from '@antfu/utils'
 import useUserStore from '../store/userStore'
-import { isArray, isArrayOfNumbers } from '../utils/array'
-import { defaultTestInit } from '../utils/baseState'
-import { getExpectedObjectProperties, user } from '../utils/dataFixtures'
+import { defaultTestInit, getExpectedObjectProperties, isArray, isArrayOfNumbers, user } from '../utils'
 
 describe('setCurrent action should return correct value', () => {
   beforeEach(() => {
