+++ test/store/customerStore.ts
@@ -1,5 +1,5 @@
 import type { UserEntity } from './Store'
-import { createPiniaEntityStore } from '.../../src/createPiniaEntityStore'
+import createPiniaEntityStore from '.../../src/createPiniaEntityStore'
 
 export const useCustomerStore = createPiniaEntityStore<UserEntity>({
   id: 'customer',
