+++ test/utils/baseState.ts
@@ -8,10 +8,3 @@ export function defaultState() {
     ...createState<UserEntity>(),
   }
 }
-
-export function defaultTestInit() {
-  const app = createApp({})
-  const pinia = createPinia()
-  app.use(pinia)
-  setActivePinia(pinia)
-}
