+++ test/utils/test.ts
@@ -1,0 +1,6 @@
+export function defaultTestInit() {
+  const app = createApp({})
+  const pinia = createPinia()
+  app.use(pinia)
+  setActivePinia(pinia)
+}
